fix coalescing tcp packets
When the pressure on the socket is high, there is not any guarantee
for one data event per packet. Indeed multiple json messages are
coalesced in one tcp packet.

We need to add a delimeter (an arbitrary '\0', compatible with C
strings, just in case) to separate the packets in valid json messages.

So we add the delimeter when the json message is stringified and we
parse the packet to split it finally in multiple json messages.

Fixes #20