add test for coalesced TCP packets
This test is related to issue #20: json messages are coalesced in one
TCP packet and json parsing fails.

Indeed, before version 1.3.5, this test fails for timeout, because
the event listener is not triggered.

Fixed using a delimeter (see PR for details): the delimeter is added
when json message is stringified and then the message is splitted by
the delimeter when it is parsed.