encapsulate message delimiter
This refactoring allows to encapsulate the arbitrary delimter
character in Message code.

Message.split separates the json messages coalesced in the TCP packet.

Notice pop call removes the last empty string in the array due to the
trailing delimiter: "1,2,3,".split(",") // [ '1', '2', '3', '' ].

It avoids also to check if event is empty in the loop.